Add files via upload
The project name is " Student Attendance System ". This project is made using the Python Kivy library. Basically python Kivy library is used for developing an Android application. For taking student attendance I used the system of QR code. When the student signs the application, In the background of the application it automatically generates the QR code of that particular student. When teachers scan this student's QR it automatically saves data of the student in the database.